backend/requirements.txt
fastapi
uvicorn
google-generativeai
python-dotenv
pydantic[email]
SQLAlchemy
pymysql
passlib[bcrypt]
python-jose[cryptography]
python-multipart
